## Description ##
This change fuses FullyConnected operator with elemwise_add after it if possible. It is done for both float and quantized path.

The change well optimize calculation on quantized graph with full quantization mode. Below are the measured results of the following command:
benchmark/python/dnnl/run.sh  benchmark/python/dnnl/fc_add.py
run before and after this PR. Measurements are done on AWS EC2 instance c6i.16xlarge (Xeon(R) Platinum 8375C CPU).

elemwise_add, float  

|    Shape    | Hidden | Before [ms] | After [ms] | Improvement |
|------------:|-------:|------------:|------------:|-----------:|
| (   1, 224) |    512 |     0.165 |     0.151 |   8%  |
| (   1, 224) |   4096 |     0.169 |     0.150 |  11%  |
| (  16,1024) |   1024 |     0.274 |     0.245 |  11%  |
| (  32,4096) |   1024 |     0.634 |     0.611 |   4%  |
| (  32,4096) |   4096 |     2.352 |     2.299 |   2%  |
| ( 512, 512) |   4096 |     1.517 |     1.414 |   7%  |


elemwise_add, mode = smart, granularity = tensor-wise  

|    Shape    | Hidden | Before [ms] | After [ms] | Improvement |
|------------:|-------:|------------:|------------:|-----------:|
| (   1, 224) |    512 |     0.182 |     0.173 |    5%  |
| (   1, 224) |   4096 |     0.184 |     0.169 |    8%  |
| (  16,1024) |   1024 |     0.246 |     0.235 |    4%  |
| (  32,4096) |   1024 |     0.328 |     0.317 |    3%  |
| (  32,4096) |   4096 |     0.573 |     0.571 |    0%  |
| ( 512, 512) |   4096 |     0.819 |     0.730 |   11%  |


elemwise_add, mode = smart, granularity = channel-wise  

|    Shape    | Hidden | Before [ms] | After [ms] | Improvement |
|------------:|-------:|------------:|------------:|-----------:|
| (   1, 224) |    512 |     0.164 |     0.138 |    16%  |
| (   1, 224) |   4096 |     0.152 |     0.143 |     6%  |
| (  16,1024) |   1024 |     0.213 |     0.199 |     7%  |
| (  32,4096) |   1024 |     0.300 |     0.285 |     5%  |
| (  32,4096) |   4096 |     0.545 |     0.542 |     1%  |
| ( 512, 512) |   4096 |     0.778 |     0.689 |    11%  |


elemwise_add, mode = full, granularity = tensor-wise  

|    Shape    | Hidden | Before* [ms] | After [ms] | Improvement |
|------------:|-------:|------------:|------------:|-----------:|
| (   1, 224) |    512 |     0.169 |     0.154 |     9%  |
| (   1, 224) |   4096 |     0.208 |     0.158 |    24%  |
| (  16,1024) |   1024 |     0.270 |     0.212 |    21%  |
| (  32,4096) |   1024 |     0.359 |     0.293 |    18%  |
| (  32,4096) |   4096 |     0.602 |     0.542 |    10%  |
| ( 512, 512) |   4096 |     0.873 |     0.652 |    25%  |


elemwise_add, mode = full, granularity = channel-wise  

|    Shape    | Hidden | Before* [ms] | After [ms] | Improvement |
|------------:|-------:|------------:|------------:|-----------:|
| (   1, 224) |    512 |     0.135 |     0.125 |    7%  |
| (   1, 224) |   4096 |     0.178 |     0.130 |   27%  |
| (  16,1024) |   1024 |     0.239 |     0.180 |   25%  |
| (  32,4096) |   1024 |     0.327 |     0.262 |   20%  |
| (  32,4096) |   4096 |     0.575 |     0.512 |   11%  |
| ( 512, 512) |   4096 |     0.852 |     0.625 |   27%  |

\* - before this PR fuzing FC with add in full quantize mode is broken, so results are taken from the first commit (https://github.com/apache/incubator-mxnet/pull/20821/commits/0c38ca7268a54664d8871890138cd021f49c4b55) of this PR which fix the issue

## Checklist ##
### Essentials ###
- [ ] PR's title starts with a category (e.g. [BUGFIX], [MODEL], [TUTORIAL], [FEATURE], [DOC], etc)
- [ ] Changes are complete (i.e. I finished coding on this PR)
- [ ] All changes have test coverage
- [ ] Code is well-documented