@anko-intel 2022-01-17
@mxnet-bot run ci [unix-cpu, windows-gpu]