@anko-intel 2022-01-18
@mxnet-bot run ci [windows-gpu]